Update Owl Carousel to avoid CSS errors
The slideshow controls do not display correctly in recent Firefox and
the bundled CSS has validation errors for some of the .owl-* selectors
like:

  Value Error : cursor hand is not a cursor value : hand

  scale(1.3,1.3) is not a transition value : scale(1.3,1.3)

This patch updates Owl Carousel to the most recent release 2.3.4 and
Animate.css to the most recent of the 3.x series before breaking
changes. I've removed the framework{,-min}.css files which bundled the
Owl and Animate CSS files together as it seems more maintainable to ship
the upstream minified CSS directly.

Also fixed a couple of minor bugs in the slide show:
- Clicking to advance the slideshow triggered twice.
- A few pixels of the right hand side of the previous image was
  displayed in the left margin.

Finally we only need to include the Owl CSS when the slideshow is
enabled in the settings.

I've done some basic testing of the slideshow functions with no obvious
regressions.