nggAutocomplete not working with jQuery 1.9 and later
The "selector" property was deprecated and then removed which means none
of the autocomplete drop-downs work anymore.

Fix using a suggestion from here:
  https://stackoverflow.com/questions/9382028/get-the-current-jquery-selector-string

In our case we only call nggAutocomplete with an id selector so we can
simplify it somewhat.

Also removed a workaround from 2005 which causes the following error:

  Uncaught TypeError: jQuery(...).zIndex is not a function

Tested the TinyMCE plugin in Firefox and Chrome.

Thanks for @feeschwer for reporting this and providing the initial
solution:
  https://wordpress.org/support/topic/mehrere-anpassung-fur-php-8/